create a common decoration material list service file.
